Refactor API constants and move shuffleArray function to util.js; add image preview functionality in form.js
