+++ js/render.js
@@ -1,5 +1,6 @@
 import { renderBigPicture } from './big-picture.js';
 import { debounce } from './util.js';
+import { shuffleArray } from './util.js';
 
 const ACTIVE_CLASS = 'img-filters__button--active';
 
@@ -11,18 +12,6 @@ const pictureSection = document.querySelector('.pictures');
 const templatePic = document.querySelector('#picture').content.querySelector('.picture');
 let pictureListFragment = document.createDocumentFragment();
 let defaultPictures = '';
-
-function shuffleArray(array) {
-  const length = array.length;
-  const shuffle = array.slice();
-  for (let i = length - 1; i > 0; i -= 1) {
-    const random = Math.floor(Math.random() * (i + 1));
-    const current = shuffle[i];
-    shuffle[i] = shuffle[random];
-    shuffle[random] = current;
-  }
-  return shuffle.slice(0, 10);
-}
 
 function sortByCommentCount(a, b) {
   return b.comments.length - a.comments.length;
