+++ js/util.js
@@ -14,4 +14,16 @@ function debounce (callback, timeoutDelay = 500) {
   };
 }
 
-export {onEscapePress, debounce};
+function shuffleArray(array) {
+  const length = array.length;
+  const shuffle = array.slice();
+  for (let i = length - 1; i > 0; i -= 1) {
+    const random = Math.floor(Math.random() * (i + 1));
+    const current = shuffle[i];
+    shuffle[i] = shuffle[random];
+    shuffle[random] = current;
+  }
+  return shuffle.slice(0, 10);
+}
+
+export {onEscapePress, debounce, shuffleArray};
